requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
django-cors-headers>=1.3.1,<=2.0.1

psycopg2>2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
python-decouple>=3.0.0,<=3.3.0
django-heroku=0.3.1
gunicorn=20.0.4

flake8>=3.6.0,<3.7.0
